+++ apply.html
@@ -21,16 +21,16 @@
                 
                 <div>
                     <label for="">Name: </label>
-                    <input id="name" type="text">
+                    <input id="name" type="text" name="name">
                 </div>
             </section>
                 
             <fieldset>
                 <legend>Are you 21?</legend>
 
-                <input id="yes" type="radio" value="yes" name="age" required>
+                <input id="yes" type="radio" value="yes" name="is21" required>
                 <label for="yes">Yes</label>
-                <input id="no" type="radio" value="no" name="age">
+                <input id="no" type="radio" value="no" name="is21">
                 <label for="no">No</label>
             </fieldset>
