+++ src/apply.js
@@ -1,5 +1,13 @@
+import getApplicant from './get-applicant.js';
+
 // reffrence needed DOM notes
 const form = document.getElementById('food-taster-application');
-// import name
+
+form.addEventListener('submit', (event) => {
+    event.preventDefault();
+    const formData = new FormData(form);
+    const applicant = getApplicant(formData);
+    console.log(applicant);
+}); 
 
 // subscribe to event, submit button
