+++ tests/tests.js
@@ -1,25 +1,12 @@
 const test = QUnit.test;
-
-function getApplicant(formData) {
-    const age = formData.get('age') === 'yes';
-    const likeFood = parseInt(formData.get('likeFood'));
-
-    const applicant = {
-        name: formData.get('name'),
-        age: age,
-        cuisine: formData.get('cuisine'),
-        meals: formData.getAll('meals'),
-        likeFood: likeFood
-    };
-    return applicant;
-}
+import getApplicant from '../src/get-applicant.js';
 
 test('return completed form in object', function(assert) {
     //Arrange
     // Set up your parameters and expectations
     const expected = {
         name: 'josh',
-        age: true,
+        is21: true,
         cuisine: 'indian',
         meals: ['breakfast', 'lunch'],
         likeFood: 10,
@@ -28,11 +15,11 @@ test('return completed form in object', function(assert) {
 
     const formData = new FormData();
     formData.set('name', expected.name);
-    formData.set('age', 'yes');
+    formData.set('is21', 'yes');
     formData.set('cuisine', expected.cuisine);
     formData.set('meals', expected.meals[0]);
     formData.append('meals', expected.meals[1]);
-    formData.set('likeFood', '10');
+    formData.set('like-food', '10');
